requirements.txt
Flask==0.10.1
Flask_RESTful==0.3.2
python-instagram==1.1.3
pymongo==2.8
redis==2.10.3
APScheduler==3.0.1
enum==0.4.4